[scripting] Remote download/update script files
Pipy now can be started with a URL pointing to a remote HTTP resource and run the script from there.
When multiple scripts are needed, point to a text file including all the required filenames coming from under the same URL.
Pipy also checks every 5s the ETag and Last-Modified headers using HEAD request to discover an update and in turn reload the scripts.