[core] Add a wrapper outside asio::steady_timer to cope with cancelation issue
After cancelling a asio::steady_timer, the handler to async_wait still has a chance to
get called if the timer has already expired, which in turn may lead to a call into
a free handler and then crash.

We add a wrapper that nullifies the handler at cancalation to cope with the issue.