[build] check before adding -Wno-delete-non-abstract-non-virtual-dtor
-Wdelete-non-abstract-non-virtual-dtor was split into two groups back
into https://reviews.llvm.org/D56405. but this change has not backported
to llvm-toolset-7, which in turn packages Clang/LLVM 5.x.

to avoid the unknown warning options warning messages from the compiler
when compiling the tree on el7, we need to check for the existence of
this warning option before using it.

in this change,

* check for the existence of the
  "-Wdelete-non-abstract-non-virtual-dtor", before disabling it.
* only apply this option on C++ source files. as strictly speaking,
  this option only applies to C++.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>