[chore] bump minimum require CMake version to 3.10
when building with CMake 4.1.1, it compained:

```
-- Detecting CXX compile features - done
CMake Error at deps/yajl-2.1.0/CMakeLists.txt:15 (CMAKE_MINIMUM_REQUIRED):
  Compatibility with CMake < 3.5 has been removed from CMake.

  Update the VERSION argument <min> value.  Or, use the <min>...<max> syntax
  to tell CMake that the project requires at least <min> but has been updated
  to work with policies introduced by <max> or earlier.

  Or, add -DCMAKE_POLICY_VERSION_MINIMUM=3.5 to try configuring anyway.
```

in order to silence this warning, let's bump the miminum required CMake
version to 3.5, which is available on major distro LTS versions. and the
parent project also use 3.10 or 3.15 as the minimum required CMake version.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>