[chore] use the new behavior of CMake policy CMP0077
when building with CMake 4.1.1, it complains:

```
CMake Warning (dev) at deps/libyaml-0.2.5/CMakeLists.txt:10 (option):
  Policy CMP0077 is not set: option() honors normal variables.  Run "cmake
  --help-policy CMP0077" for policy details.  Use the cmake_policy command to
  set the policy and suppress this warning.

  For compatibility with older versions of CMake, option is clearing the
  normal variable 'BUILD_SHARED_LIBS'.
This warning is for project developers.  Use -Wno-dev to suppress it.
```

as libyaml is embeded in the parent project, and it prefers to build a
static library. and we want to build static libraries as set in the
parent's CMakeLists.txt. the new behavior which does nothing when a
normal variable of the same name exists. this is exactly what we expect.

so let's use the new behavior if the policy is set by the used CMake.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>