[chore] change to build with CMake 4.1.1 (#209)
* [chore] bump minimum require CMake version to 3.10

when building with CMake 4.1.1, it compained:

```
-- Detecting CXX compile features - done
CMake Error at deps/yajl-2.1.0/CMakeLists.txt:15 (CMAKE_MINIMUM_REQUIRED):
  Compatibility with CMake < 3.5 has been removed from CMake.

  Update the VERSION argument <min> value.  Or, use the <min>...<max> syntax
  to tell CMake that the project requires at least <min> but has been updated
  to work with policies introduced by <max> or earlier.

  Or, add -DCMAKE_POLICY_VERSION_MINIMUM=3.5 to try configuring anyway.
```

in order to silence this warning, let's bump the miminum required CMake
version to 3.5, which is available on major distro LTS versions. and the
parent project also use 3.10 or 3.15 as the minimum required CMake version.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>

* [chore] use the new behavior of CMake policy CMP0077

when building with CMake 4.1.1, it complains:

```
CMake Warning (dev) at deps/libyaml-0.2.5/CMakeLists.txt:10 (option):
  Policy CMP0077 is not set: option() honors normal variables.  Run "cmake
  --help-policy CMP0077" for policy details.  Use the cmake_policy command to
  set the policy and suppress this warning.

  For compatibility with older versions of CMake, option is clearing the
  normal variable 'BUILD_SHARED_LIBS'.
This warning is for project developers.  Use -Wno-dev to suppress it.
```

as libyaml is embeded in the parent project, and it prefers to build a
static library. and we want to build static libraries as set in the
parent's CMakeLists.txt. the new behavior which does nothing when a
normal variable of the same name exists. this is exactly what we expect.

so let's use the new behavior if the policy is set by the used CMake.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>

* [chore] use EXECUTE_PROCESS() instead of EXEC_PROGRAM() in deps/yajl

according to https://cmake.org/cmake/help/latest/policy/CMP0153.html,
exec_program() should not be called. let's switch to the recommended
execute_process().

Signed-off-by: Kefu Chai <tchaikov@gmail.com>

---------

Signed-off-by: Kefu Chai <tchaikov@gmail.com>